refactor(scripts): modularize script files
