+++ hathor/transaction/scripts/__init__.py
@@ -1,0 +1,40 @@
+#  Copyright 2023 Hathor Labs
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#  http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+
+from hathor.transaction.scripts.construct import (
+    create_base_script,
+    create_output_script,
+    get_sigops_count,
+    parse_address_script,
+)
+from hathor.transaction.scripts.execute import ScriptExtras, script_eval
+from hathor.transaction.scripts.hathor_script import HathorScript
+from hathor.transaction.scripts.multi_sig import MultiSig
+from hathor.transaction.scripts.nano_contract_match_values import NanoContractMatchValues
+from hathor.transaction.scripts.opcode import Opcode
+from hathor.transaction.scripts.p2pkh import P2PKH
+
+__all__ = [
+    'Opcode',
+    'P2PKH',
+    'MultiSig',
+    'NanoContractMatchValues',
+    'HathorScript',
+    'ScriptExtras',
+    'parse_address_script',
+    'create_base_script',
+    'create_output_script',
+    'script_eval',
+    'get_sigops_count',
+]
